+++ aiworkspace-20210204/alibabacloud_aiworkspace20210204/models/_get_permission_shrink_request.py
@@ -8,16 +8,21 @@ class GetPermissionShrinkRequest(DaraModel):
     def __init__(
         self,
         accessibility: str = None,
+        caller_type: str = None,
+        caller_uid: str = None,
         creator: str = None,
         labels_shrink: str = None,
         option: str = None,
         resource: str = None,
+        security_token: str = None,
     ):
         # The accessibility. Valid values:
         # 
         # *   PUBLIC: All members in the workspace can access the workspace.
         # *   PRIVATE: Only the creator can access the workspace.
         self.accessibility = accessibility
+        self.caller_type = caller_type
+        self.caller_uid = caller_uid
         # The UID of the Alibaba Cloud account that is used to create the workspace.
         self.creator = creator
         self.labels_shrink = labels_shrink
@@ -28,6 +33,7 @@ class GetPermissionShrinkRequest(DaraModel):
         self.option = option
         # The resource.
         self.resource = resource
+        self.security_token = security_token
 
     def validate(self):
         pass
@@ -39,6 +45,12 @@ class GetPermissionShrinkRequest(DaraModel):
             result = _map
         if self.accessibility is not None:
             result['Accessibility'] = self.accessibility
+
+        if self.caller_type is not None:
+            result['CallerType'] = self.caller_type
+
+        if self.caller_uid is not None:
+            result['CallerUid'] = self.caller_uid
 
         if self.creator is not None:
             result['Creator'] = self.creator
@@ -52,12 +64,21 @@ class GetPermissionShrinkRequest(DaraModel):
         if self.resource is not None:
             result['Resource'] = self.resource
 
+        if self.security_token is not None:
+            result['SecurityToken'] = self.security_token
+
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
         if m.get('Accessibility') is not None:
             self.accessibility = m.get('Accessibility')
+
+        if m.get('CallerType') is not None:
+            self.caller_type = m.get('CallerType')
+
+        if m.get('CallerUid') is not None:
+            self.caller_uid = m.get('CallerUid')
 
         if m.get('Creator') is not None:
             self.creator = m.get('Creator')
@@ -70,6 +91,9 @@ class GetPermissionShrinkRequest(DaraModel):
 
         if m.get('Resource') is not None:
             self.resource = m.get('Resource')
+
+        if m.get('SecurityToken') is not None:
+            self.security_token = m.get('SecurityToken')
 
         return self